Fix placeholders that didn't display
The problem was the useTranslation hook was always debounced,
so it only returned the translation for the last placeholder.

The solution was to split the hook into normal and debounced versions.
The normal version is used for getting placeholder translations.
The debounced version is used for getting user input translations.